Refactor and test commonChars
